% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chapter1.r
\name{Commutator_Kmn}
\alias{Commutator_Kmn}
\title{Commutator_Kmn}
\usage{
Commutator_Kmn(m, n)
}
\arguments{
\item{m}{Row-dimension}

\item{n}{Col-dimension}
}
\value{
An commutation matrix matrix of dimension \eqn{mn x mn}
}
\description{
Provides the commutation matrix
}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021 (p.7, (1.12)).
}
