% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commutators.R
\name{Commutator_Kperm}
\alias{Commutator_Kperm}
\title{Commutator_Kperm}
\usage{
Commutator_Kperm(perm, dims)
}
\arguments{
\item{perm}{vector indicating the permutation of the order in the kronecker product}

\item{dims}{vector indicating the dimensions of the vectors}
}
\value{
A permutation matrix
}
\description{
Produces any permutation of kronecker products of vectors of any length
}
\examples{
dims=c(2,3,2)
perm = c(1,3,2)
K.matr(perm,dims)
perm =c(3,1,4,2) dims=4 # All vectors with dimension 4
}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.
}
