% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commutators.R
\name{Commutator_Mixing}
\alias{Commutator_Mixing}
\title{Commutator_Mixing}
\usage{
Commutator_Mixing(d1, d2)
}
\arguments{
\item{d1}{dimension of the first group of vectors}

\item{d2}{dimension of the second group of vectors}
}
\description{
mixing commutator
}
\examples{
d1 <- c(2, 3, 2)
d2<-  c(3 ,2, 2)
Commutator_Mixing(d1,d2)
}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021. (4.58) p. 171
}
