% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomCums.r
\name{Cum2MomMulti}
\alias{Cum2MomMulti}
\title{Get moments from  cumulants (multivariate)}
\usage{
Cum2MomMulti(cum)
}
\arguments{
\item{cum}{the list of n d-variate cumulants in vector form starting from the fist cumulant -
the vector of means and arriving to the n-th order d-variate cumulant in vector form}
}
\value{
Mom  the list of n vectors of d-variate moments
}
\description{
Obtains a vector of d-variate moments from a vector of
d-variate cumulants
}
\examples{
#\code{cum} contains the  vector cumulants up to the fifth order of the bivariate
#standard normal distribution
cum<-list(c(0,0),c(1,0,0,1),c(0,0,0,0,0,0,0,0),c(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0),
c(rep(0,32)))
Cum2MomMulti(cum)

}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 3.4 formula 3.40
}
\seealso{
Other Moments and cumulants: 
\code{\link{Cum2Mom}()},
\code{\link{Mom2CumMulti}()},
\code{\link{Mom2Cum}()}
}
\concept{Moments and cumulants}
