% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{distr_Uni_Rand}
\alias{distr_Uni_Rand}
\title{Random Uniform on the sphere}
\usage{
distr_Uni_Rand(n, d)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}
}
\value{
A random matrix \eqn{n \times d}
}
\description{
Generate random d-vectors from the Uniform distribution on the sphere
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021

S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644.
}
\seealso{
Other Random generation: 
\code{\link{distr_CFUSN_Rand}()},
\code{\link{distr_CFUSSD_Rand}()},
\code{\link{distr_SkewNorm_Rand}()}

Other Multivariate distributions: 
\code{\link{distr_CFUSN_MomCum_Th}()},
\code{\link{distr_CFUSN_Rand}()},
\code{\link{distr_CFUSSD_Rand}()},
\code{\link{distr_SkewNorm_EVSK_Th}()},
\code{\link{distr_SkewNorm_MomCum_Th}()},
\code{\link{distr_SkewNorm_Rand}()},
\code{\link{distr_UniAbs_EVSK_Th}()},
\code{\link{distr_Uni_EVSK_Th}()},
\code{\link{distr_Uni_MomCum_Th}()},
\code{\link{distr_ZabsM_MomCum_Th}()},
\code{\link{distr_Zabs_MomCum_Th}()}
}
\concept{Multivariate distributions}
\concept{Random generation}
