% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Gram_Charlier}
\alias{Esti_Gram_Charlier}
\title{Gram-Charlier approximation to a multivariate density}
\usage{
Esti_Gram_Charlier(X, k = 4, cum = NULL)
}
\arguments{
\item{X}{A matrix of d-variate data}

\item{k}{the order of the approximation, by default set to 4; 
(k must not be smaller than 3 or greater than 8)}

\item{cum}{if NULL (default) the cumulant vector is estimated from X. 
If \code{cum} is provided no estimation of cumulants is performed.}
}
\value{
The vector of the Gram-Charlier density evaluated at X
}
\description{
Provides the truncated Gram-Charlier approximation to a multivariate density. Approximation can
be up to the first k=8 cumulants.
}
\examples{
# Gram-Charlier density approximation (k=4) of data generated from 
# a bivariate skew-gaussian distribution
n<-50
alpha<-c(10,0)
omega<-diag(2)
X<-distr_SkewNorm_Rand(n,omega,alpha)
EC<-Esti_EVSK(X)
fy4<-Esti_Gram_Charlier(X[1:5,],cum=EC)
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021. Section 4.7.
}
