% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_H3}
\alias{Esti_H3}
\title{Estimate the third d-variate polynomial}
\usage{
Esti_H3(x)
}
\arguments{
\item{x}{a d-variate data vector}
}
\description{
The vector x is standardized and the skewness vector is computed
}
\examples{
x<-mvtnorm::rmvnorm(100,rep(0,3))
H3<-Esti_H3(x)
}
