% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_H4}
\alias{Esti_H4}
\title{Estimate the fourth d-variate polynomial}
\usage{
Esti_H4(x)
}
\arguments{
\item{x}{a d-variate data vector}
}
\description{
The vector x is standardized and the kurtosis vector is computed
}
\examples{
x<-mvtnorm::rmvnorm(100,rep(0,3))
H4<-Esti_H4(x)
}
