% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Hermite_Poly_HN_Multi}
\alias{Esti_Hermite_Poly_HN_Multi}
\title{Estimate the N-th d-variate Hermite polynomial}
\usage{
Esti_Hermite_Poly_HN_Multi(x, N)
}
\arguments{
\item{x}{a d-variate data vector}

\item{N}{the order of the d-variate Hermite polynomial}
}
\value{
The vector of the N-th d-variate polynomial
}
\description{
The vector x is standardized and the N-th d-variate polynomial is computed
}
\examples{
x<-MASS::mvrnorm(100,rep(0,3),diag(3))
H3<-Esti_Hermite_Poly_HN_Multi(x,3)
}
