% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Kurt_Mardia}
\alias{Esti_Kurt_Mardia}
\title{Mardia Kurtosis Index}
\usage{
Esti_Kurt_Mardia(x)
}
\arguments{
\item{x}{A vector of multivariate data}
}
\value{
The kurtosis index
}
\description{
Mardia Kurtosis Index
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.1
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mori}()},
\code{\link{Esti_SkewInd_Mori}()},
\code{\link{Esti_SkewVec_Mori}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Kurt_Mardia}()},
\code{\link{Kurt_Mori}()},
\code{\link{Skew_Mardia}()},
\code{\link{Skew_Mori}()}
}
\concept{Indexes}
