% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Kurt_Mori}
\alias{Esti_Kurt_Mori}
\title{Mori's kurtosis matrix}
\usage{
Esti_Kurt_Mori(x)
}
\arguments{
\item{x}{A vector of multivariate data}
}
\value{
The kurtosis matrix
}
\description{
Mori's kurtosis matrix
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.9
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_SkewInd_Mori}()},
\code{\link{Esti_SkewVec_Mori}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Kurt_Mardia}()},
\code{\link{Kurt_Mori}()},
\code{\link{Skew_Mardia}()},
\code{\link{Skew_Mori}()}
}
\concept{Indexes}
