% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Kurt_MRSz}
\alias{Esti_Kurt_MRSz}
\title{Estimation of  Mori,  Rohatgi, Szekely (CMRSz's) kurtosis matrix}
\usage{
Esti_Kurt_MRSz(x)
}
\arguments{
\item{x}{A matrix of multivariate data}
}
\value{
\code{MRSz.Kurtosis} The kurtosis matrix

\code{p.value} The p-value for the hypothesis of symmetry under the
Gaussian assumption
}
\description{
Estimation of  Mori,  Rohatgi, Szekely (CMRSz's) kurtosis matrix
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.9
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Kurt_Total}()},
\code{\link{Esti_Skew_MRSz}()},
\code{\link{Esti_Skew_Mardia}()}
}
\concept{Indexes}
