% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Kurt_Total}
\alias{Esti_Kurt_Total}
\title{Estimation of the Total Kurtosis Index}
\usage{
Esti_Kurt_Total(x)
}
\arguments{
\item{x}{A matrix of multivariate data}
}
\value{
\code{Total.Kurtosis} The total kurtosis index

\code{p.value} The p-value under the Gaussian hypothesis
}
\description{
Estimation of the Total Kurtosis Index
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.1
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_CMRSz}()},
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Skew_MRSz}()},
\code{\link{Esti_Skew_Mardia}()}
}
\concept{Indexes}
