% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Kurt_Variance_Th}
\alias{Esti_Kurt_Variance_Th}
\title{Asymptotic Variance of the estimated  kurtosis vector}
\usage{
Esti_Kurt_Variance_Th(cum)
}
\arguments{
\item{cum}{The theoretical/estimated cumulants up to the 8th order in vector form}
}
\value{
The matrix of theoretical/estimated variance
}
\description{
Warning: the function requires 8! computations, for d>3, the timing required maybe large.
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021. Ch. 6, formula (6.26)
}
\seealso{
Other Estimation: 
\code{\link{Esti_EVSK}()},
\code{\link{Esti_MMom_MCum}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Esti_Skew_Variance_Th}()}
}
\concept{Estimation}
