% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_MMom_MCum}
\alias{Esti_MMom_MCum}
\title{Estimate for Multivariate Moments and Cumulants}
\usage{
Esti_MMom_MCum(X, k, centr = 0, sc_sigM = 1)
}
\arguments{
\item{X}{d-vector data}

\item{k}{The highest moment order}

\item{centr}{set to 1 (and sc_sigM=0) if only centering is needed}

\item{sc_sigM}{set to 1 (and centr=0) if standardization of multivariate data is needed}
}
\description{
Provides estimates of multivariate moments and cumulants up to order k.
By default data are standardized, only centering can be used
}
\seealso{
Other Estimation: 
\code{\link{Esti_MVSK}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Esti_Variance_MVSK}()},
\code{\link{Variance_of_Esti_Kurt_Th}()},
\code{\link{Variance_of_Esti_Skew_Th}()}
}
\concept{Estimation}
