% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_MVSK}
\alias{Esti_MVSK}
\title{Estimates for multivariate Skewness and Kurtosis}
\usage{
Esti_MVSK(X)
}
\arguments{
\item{X}{d-variate data vector}
}
\value{
The list of the estimated quantities
}
\description{
Provides estimates of mean, variance, skewness and kurtosis vectors for  d-variate data
}
\examples{
x<-mvtnorm::rmvnorm(100,rep(0,d),sigma = 3*diag(rep(1,d)))
MVSK<-Esti_MVSK(x)
names(MVSK)
MVSK$estSkew
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.
}
\seealso{
Other Estimation: 
\code{\link{Esti_MMom_MCum}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Esti_Variance_MVSK}()},
\code{\link{Variance_of_Esti_Kurt_Th}()},
\code{\link{Variance_of_Esti_Skew_Th}()}
}
\concept{Estimation}
