% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Skew_Mardia}
\alias{Esti_Skew_Mardia}
\title{Mardia Skewness index}
\usage{
Esti_Skew_Mardia(x)
}
\arguments{
\item{x}{A vector of multivariate data}
}
\value{
The skewness index

The p-value
}
\description{
Compute the multivariate Mardia's skewness index and
provides the p-value for the hypothesis of zero symmetry under the
Gaussian assumption
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.1
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Kurt_Mori}()},
\code{\link{Esti_SkewInd_Mori}()},
\code{\link{Esti_SkewVec_Mori}()},
\code{\link{Kurt_Mardia}()},
\code{\link{Kurt_Mori}()},
\code{\link{Skew_Mardia}()},
\code{\link{Skew_Mori}()}

Other Estimation: 
\code{\link{Esti_MMom_MCum}()},
\code{\link{Esti_MVSK}()},
\code{\link{Esti_Variance_MVSK}()},
\code{\link{Variance_of_Esti_Kurt_Th}()},
\code{\link{Variance_of_Esti_Skew_Th}()}
}
\concept{Estimation}
\concept{Indexes}
