% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Skew_MRSz}
\alias{Esti_Skew_MRSz}
\title{Estimation of Mori, Rohatgi, Szekely (MRSz's) skewness vector}
\usage{
Esti_Skew_MRSz(x)
}
\arguments{
\item{x}{A matrix of multivariate data}
}
\value{
\code{MRSz.Skewness.Vector} The skewness vector

\code{MRSz.Skewness.Index} The skewness index

\code{p.value} The p-value for the hypothesis of symmetry under the
Gaussian assumption
}
\description{
Estimation of Mori, Rohatgi, Szekely (MRSz's) skewness vector
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.2

S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_CMRSz}()},
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Kurt_Total}()},
\code{\link{Esti_Skew_Mardia}()}
}
\concept{Indexes}
