% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_SkewInd_Mori}
\alias{Esti_SkewInd_Mori}
\title{Mori's skewness index}
\usage{
Esti_SkewInd_Mori(x)
}
\arguments{
\item{x}{A vector of multivariate data}
}
\value{
The skewness index

The p-value for the hypothesis of zero symmetry under the
Gaussian assumption
}
\description{
Mori's skewness index
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Example 6.2
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Kurt_Mori}()},
\code{\link{Esti_SkewVec_Mori}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Kurt_Mardia}()},
\code{\link{Kurt_Mori}()},
\code{\link{Skew_Mardia}()},
\code{\link{Skew_Mori}()}
}
\concept{Indexes}
