% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Esti_Variance_MVSK}
\alias{Esti_Variance_MVSK}
\title{Estimated Variance of  skewness and kurtosis}
\usage{
Esti_Variance_MVSK(X)
}
\arguments{
\item{X}{d-variate data}
}
\value{
The list ov covaraince matrices of the skewness and kurtosis vectors
}
\description{
Provides the estimated covariance matrices of the data-estimated skewness and kurtosis
vectors.
}
\examples{
x<-mvtnorm::rmvnorm(100,rep(0,d),sigma = 3*diag(rep(1,d)))
MVSK<-Esti_MVSK(x)
EV_MVSK<-Esti_Variance_MVSK(x)
names(EV_MVSK)
EV_MVSK$Vari_Skew_e
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Chapter 6, formulae (6.13) and (6.22).
}
\seealso{
Other Estimation: 
\code{\link{Esti_MMom_MCum}()},
\code{\link{Esti_MVSK}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Variance_of_Esti_Kurt_Th}()},
\code{\link{Variance_of_Esti_Skew_Th}()}
}
\concept{Estimation}
