% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_CoeffMulti}
\alias{Hermite_CoeffMulti}
\title{Multivariate Hermite coefficients}
\usage{
Hermite_CoeffMulti(N, d)
}
\description{
Multivariate Hermite coefficients
}
\seealso{
Other Hermite: 
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Inv}()},
\code{\link{Hermite_Poly_NH_Multi_Inv}()}
}
\concept{Hermite}
