% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_Poly_HN}
\alias{Hermite_Poly_HN}
\title{Hermite_Poly_HN
% (4.24) p.157
given x and sigma2=sigma^2
output values of  Hermite polynomials from 1 to N}
\usage{
Hermite_Poly_HN(x, N, sigma2 = 1)
}
\description{
Hermite_Poly_HN
% (4.24) p.157
given x and sigma2=sigma^2
output values of  Hermite polynomials from 1 to N
}
\seealso{
Other Hermite: 
\code{\link{Hermite_CoeffMulti}()},
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_NH_Inv}()},
\code{\link{Hermite_Poly_NH_Multi_Inv}()}
}
\concept{Hermite}
