% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_Poly_HN_Multi}
\alias{Hermite_Poly_HN_Multi}
\title{Multivariate Hermite Polynomials}
\usage{
Hermite_Poly_HN_Multi(x, Sig2, N)
}
\description{
Multivariate Hermite Polynomials
}
\examples{
x<-c(1,3)  
N<-3
Sig2<-matrix(c(1,0,0,1),2,2,byrow = T)
Hermite_Poly_HN_Multi(x,N,Sig2)
}
\seealso{
Other Hermite: 
\code{\link{Hermite_CoeffMulti}()},
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Inv}()},
\code{\link{Hermite_Poly_NH_Multi_Inv}()}
}
\concept{Hermite}
