% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_Poly_NH_Inv}
\alias{Hermite_Poly_NH_Inv}
\title{Hermite_Poly_HN_Inv
% (4.23) p.157
given Hermite polynomials  H_N_x from 1 to N and sigma2=sigma^2
output powers of x_val: x^n, n=1:N}
\usage{
Hermite_Poly_NH_Inv(H_N_x, sigma2 = 1)
}
\description{
Hermite_Poly_HN_Inv
% (4.23) p.157
given Hermite polynomials  H_N_x from 1 to N and sigma2=sigma^2
output powers of x_val: x^n, n=1:N
}
\seealso{
Other Hermite: 
\code{\link{Hermite_CoeffMulti}()},
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Multi_Inv}()}
}
\concept{Hermite}
