% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_Poly_NH_Multi_Inv}
\alias{Hermite_Poly_NH_Multi_Inv}
\title{Inverse of d-variate T-Hermite Polynomial}
\usage{
Hermite_Poly_NH_Multi_Inv(H_N_X, Sig2 = diag(length(x)), N)
}
\arguments{
\item{Sig2}{The variance matrix of x, the default is set to unit matrix}

\item{H_N_x}{The list  of  d-variate T-Hermite Polynomials of order  from 1 to N evaluated at x}
}
\value{
The list of \eqn{x}, \eqn{x^{\otimes 2)}},... \eqn{x^{\otimes N}}
}
\description{
Compute the powers of vector variate x when Hermite polynomials are given
}
\examples{
Sig2=matrix(c(1,0,0,1),2,2,byrow=T)
x<-c(1,3)
N<-4
H_N_X<-Hermite_Poly_HN_Multi(x,Sig2,N)
x_ad_n <- Hermite_Poly_NH_Multi_Inv(H_N_X,Sig2,N)
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 4.6.2, (4.72), p.223
}
\seealso{
Other Hermite: 
\code{\link{Hermite_CoeffMulti}()},
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_N_Cov_X1_X2}()},
\code{\link{Hermite_Nth}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Inv}()}
}
\concept{Hermite}
