% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Kurt_Mardia}
\alias{Kurt_Mardia}
\title{Mardia Kurtosis index}
\usage{
Kurt_Mardia(H4)
}
\arguments{
\item{H4}{the kurtosis vector.}
}
\value{
Mardia's kurtosis index
}
\description{
Computes Mardia's kurtosis index given the kurtosis vector.
It can be used to compute estimated or population values.
}
\examples{
#Compute Mardia's kurtosis index for the multivariate Skew-Normal
alpha<-c(10,5)
omega<-diag(rep(1,2))
H4SN<-MVSK_SkewNorm_Th(omega, alpha)$KurtX
Kurt_Mardia(H4SN)
}
\references{
S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644, Example 7.
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Kurt_Mori}()},
\code{\link{Esti_SkewInd_Mori}()},
\code{\link{Esti_SkewVec_Mori}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Kurt_Mori}()},
\code{\link{Skew_Mardia}()},
\code{\link{Skew_Mori}()}
}
\concept{Indexes}
