% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Kurt_Mori}
\alias{Kurt_Mori}
\title{Mori's kurtosis matrix}
\usage{
Kurt_Mori(H4)
}
\arguments{
\item{H4}{the d-variate kurtosis vector.}
}
\value{
Mori's kurtosis matrix
}
\description{
Computes Mori's kurtosis matrix given the kurtosis vector.
It can be used to compute estimated or population values.
}
\examples{
#Compute Mori's kurtosis matrix for the multivariate Skew-Normal
alpha<-c(10,5)
omega<-diag(rep(1,2))
H4SN<-MVSK_SkewNorm_Th(omega, alpha)$KurtX
Kurt_Mori(H4SN)
}
\references{
S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644, Example 9.
}
\seealso{
Other Indexes: 
\code{\link{Esti_Kurt_Mardia}()},
\code{\link{Esti_Kurt_Mori}()},
\code{\link{Esti_SkewInd_Mori}()},
\code{\link{Esti_SkewVec_Mori}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Kurt_Mardia}()},
\code{\link{Skew_Mardia}()},
\code{\link{Skew_Mori}()}
}
\concept{Indexes}
