% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chapter1.r
\name{Matrix_Elimination}
\alias{Matrix_Elimination}
\title{Elimination Matrix}
\usage{
Matrix_Elimination(d, q)
}
\arguments{
\item{d}{dimension of a vector x}

\item{q}{power of the Kronecker product}
}
\value{
Elimination Matrix of order  ŋ\eqn{_{d,q}x d^q}, see (1.30), p.15 # \eqn{_{d,q}}
}
\description{
Eliminated the duplicated/q-plicated elements in the vector
}
\references{
Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.
Section  1.3.2 Multi-Indexing, Elimination, and Duplication, p.21,(1.32)
}
\seealso{
Other Matrix: 
\code{\link{Matrix_Qplication}()},
\code{\link{Matrix_Symmetry}()}
}
\concept{Matrix}
