% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chapter1.r
\name{Matrix_Qplication}
\alias{Matrix_Qplication}
\title{Qplication Matrix}
\usage{
Matrix_Qplication(d, q)
}
\arguments{
\item{d}{dimension of a vector x}

\item{q}{power of the Kronecker product}
}
\value{
Qplication Matrix of order \eqn{d^q x ŋ_{d,q}}, see (1.30), p.15
}
\description{
Restores the duplicated/q-plicated  elements which are eliminated
by  Matrix_Elimination in the vector,
}
\references{
Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021, p.21, (1.31)
}
\seealso{
Other Matrix: 
\code{\link{Matrix_Elimination}()},
\code{\link{Matrix_Symmetry}()}
}
\concept{Matrix}
