% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chapter1.r
\name{Matrix_Symmetry}
\alias{Matrix_Symmetry}
\title{Symmetrizer Matrix}
\usage{
Matrix_Symmetry(d, q)
}
\arguments{
\item{d}{dimension of a vector x}

\item{q}{power of the Kronecker product}
}
\value{
Symmetrizer Matrix with order d^q x d^q
}
\description{
It does not use sparse matrices or parallel computations
}
\details{
@references Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.Section 1.3.1 Symmetrization, p.14. (1.29)
}
\seealso{
Other Matrix: 
\code{\link{Matrix_Elimination}()},
\code{\link{Matrix_Qplication}()}
}
\concept{Matrix}
