% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomCums.r
\name{Mom2CumMulti}
\alias{Mom2CumMulti}
\title{Get  moments from  cumulants (multivariate)}
\usage{
Mom2CumMulti(mu)
}
\arguments{
\item{mu}{the list of n d-variate moments in vector form starting from the fist moment -
the vector of means and arriving to the n-th order d-variate moment in vector form}
}
\value{
Cum the list of n vectors of d-variate cumulants
}
\description{
Obtains a vector of d-variate cumulants from a vector of
d-variate moments
}
\examples{
\code{mu} contains the  vector moments up to the fifth orer of the bivariate
standard normal distribution
mu<-list(c(0,0),c(1,0,0,1),c(0,0,0,0,0,0,0,0),c(3,0,0,1,0,1,1,0,0,1,1,0,1,0,0,3),
c(rep(0,32)))
Mom2CumMulti(mu)

}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 3.4 formula 3.29
}
\seealso{
Other Moments and cumulants: 
\code{\link{Cum2MomMulti}()},
\code{\link{Cum2Mom}()},
\code{\link{Mom2Cum}()}
}
\concept{Moments and cumulants}
