% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MomCum_CFUSN_Th}
\alias{MomCum_CFUSN_Th}
\title{Moments and cumulants CFUSN}
\usage{
MomCum_CFUSN_Th(r, d, p, Delta)
}
\arguments{
\item{r}{The highest cumulant order}

\item{d}{The multivariate dimension and number of rows of the skewness matrix Delta}

\item{p}{The number of cols of the skewness matrix Delta}

\item{Delta}{The skewness matrix}
}
\value{
The list of theoretical  cumulants in vector form
}
\description{
Provides the theoretical cumulants of the multivariate Canonical Fundamental
Skew Normal distribution
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021, Lemma 5.3 p.251
}
\seealso{
Other Theoretical Moments, Cumulants: 
\code{\link{MVSK_SkewNorm_Th}()},
\code{\link{MVSK_UniDistr_Th}()},
\code{\link{MomCum_MZabs_Th}()},
\code{\link{MomCum_SkewNorm_Th}()},
\code{\link{MomCum_UniAbsDistr_Th}()},
\code{\link{MomCum_UniDistr_Th}()}
}
\concept{Theoretical Moments, Cumulants}
