% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MomCum_MZabs_Th}
\alias{MomCum_MZabs_Th}
\title{Moments and cumulants multivariate central folded Normal distribution}
\usage{
MomCum_MZabs_Th(r, d)
}
\arguments{
\item{r}{The highest moment (cumulant) order}

\item{d}{dimension}
}
\value{
The list of moments and cumulants in vector form
}
\description{
Provides the theoretical moments and cumulants of the multivariate central Folded
Normal distribution. By default only moments are provided.
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021, Lemma 5.2 p. 249
}
\seealso{
Other Theoretical Moments, Cumulants: 
\code{\link{MVSK_SkewNorm_Th}()},
\code{\link{MVSK_UniDistr_Th}()},
\code{\link{MomCum_CFUSN_Th}()},
\code{\link{MomCum_SkewNorm_Th}()},
\code{\link{MomCum_UniAbsDistr_Th}()},
\code{\link{MomCum_UniDistr_Th}()}
}
\concept{Theoretical Moments, Cumulants}
