% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MomCum_UniAbsDistr_Th}
\alias{MomCum_UniAbsDistr_Th}
\title{Moments of the modulus of the Uniform distribution on the sphere}
\usage{
MomCum_UniAbsDistr_Th(d)
}
\arguments{
\item{d}{vector-dimension}
}
\value{
The list of the first four moments in vector form
}
\description{
Moments (up to the 4th order) of the modulus of the d-variate Uniform distribution on
the sphere on (d-1)
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021, Lemma 5.12 p.298
}
\seealso{
Other Theoretical Moments, Cumulants: 
\code{\link{MVSK_SkewNorm_Th}()},
\code{\link{MVSK_UniDistr_Th}()},
\code{\link{MomCum_CFUSN_Th}()},
\code{\link{MomCum_MZabs_Th}()},
\code{\link{MomCum_SkewNorm_Th}()},
\code{\link{MomCum_UniDistr_Th}()}
}
\concept{Theoretical Moments, Cumulants}
