% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MomCum_UniDistr_Th}
\alias{MomCum_UniDistr_Th}
\title{Moments and cumulants Uniform Distribution on the Sphere}
\usage{
MomCum_UniDistr_Th(r, d, nCum = 0)
}
\arguments{
\item{r}{highest order of moments and cumulants}

\item{d}{dimension}
}
\value{
The list of moments and cumulants in vector form
}
\description{
By default, only moments are provided
}
\examples{
# The first four moments for d=3
MomCum_UniDistr_Th(4,3,nCum=0)
# The first four moments and cumulants for d=3
MomCum_UniDistr_Th(4,3,nCum=4)
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021 Proposition 5.3 p.297
}
\seealso{
Other Theoretical Moments, Cumulants: 
\code{\link{MVSK_SkewNorm_Th}()},
\code{\link{MVSK_UniDistr_Th}()},
\code{\link{MomCum_CFUSN_Th}()},
\code{\link{MomCum_MZabs_Th}()},
\code{\link{MomCum_SkewNorm_Th}()},
\code{\link{MomCum_UniAbsDistr_Th}()}
}
\concept{Theoretical Moments, Cumulants}
