% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MomCum_Zabs}
\alias{MomCum_Zabs}
\title{Moments and cumulants Central folded Normal distribution}
\usage{
MomCum_Zabs(n, nCum = FALSE)
}
\arguments{
\item{n}{The highest moment (cumulant) order}
}
\value{
The list of moments and cumulants
}
\description{
Provides the theoretical moments and cumulants of the univariate Central Folded
Normal distribution. By default only moments are provided.
}
\examples{
# The first three moments
MomCum_Zabs(3, nCum = F)
# The first three moments and cumulants
MomCum_Zabs(3, nCum = T)
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021, Proposition 5.1 p.242 and  formula: p. 301
}
