% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MVSK_SkewNorm_Th}
\alias{MVSK_SkewNorm_Th}
\title{MVSK multivariate Skew Normal}
\usage{
MVSK_SkewNorm_Th(Om, alf)
}
\arguments{
\item{Om}{A \eqn{d \times d} correlation matrix}

\item{alf}{shape parameter d-vector}
}
\value{
A list of theoretical values
}
\description{
Computes the theoretical values of the mean vector,  covariance, skewness vector,
total skenwness, kurtosis vector and total
kurtosis for the multivariate Skew Normal distribution
}
\examples{
alpha<-c(10,5,0)
omega<-diag(rep(1,3))
MVSK_SkewNorm_Th(omega,alpha)
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021 (5.5) p.247

S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644.
}
\seealso{
Other Theoretical Moments, Cumulants: 
\code{\link{MVSK_UniDistr_Th}()},
\code{\link{MomCum_CFUSN_Th}()},
\code{\link{MomCum_MZabs_Th}()},
\code{\link{MomCum_SkewNorm_Th}()},
\code{\link{MomCum_UniAbsDistr_Th}()},
\code{\link{MomCum_UniDistr_Th}()}
}
\concept{Theoretical Moments, Cumulants}
