% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{MVSK_UniDistr_Th}
\alias{MVSK_UniDistr_Th}
\title{MVSK Uniform on the sphere}
\usage{
MVSK_UniDistr_Th(d, nCum = 0)
}
\arguments{
\item{d}{dimensions}
}
\description{
Computes the theoretical values of the mean vector,  covariance, skewness vector,
total skenwness, kurtosis vector and total
kurtosis for the Uniform distribution on the sphere.  Note that Skewness is ZERO
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021 Proposition 5.3 p.297

S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644.
}
\seealso{
Other Theoretical Moments, Cumulants: 
\code{\link{MVSK_SkewNorm_Th}()},
\code{\link{MomCum_CFUSN_Th}()},
\code{\link{MomCum_MZabs_Th}()},
\code{\link{MomCum_SkewNorm_Th}()},
\code{\link{MomCum_UniAbsDistr_Th}()},
\code{\link{MomCum_UniDistr_Th}()}
}
\concept{Theoretical Moments, Cumulants}
