% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_2Perm}
\alias{Partition_2Perm}
\title{Permute elements according to partition pU}
\usage{
Partition_2Perm(pU)
}
\arguments{
\item{pU}{A partition matrix generated by Partition_All}
}
\value{
perm_pU The elements \code{1} to \code{N} permuted according to \code{pU}
}
\description{
Permute elements according to partition pU
}
\examples{
PA<-Partition_All(4)
Partition_2Perm(PA$U[[3]])

}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 1.4.4
}
\seealso{
Other Partitions: 
\code{\link{Partition_All}()},
\code{\link{Partition_Generator}()},
\code{\link{Partition_Indecomposable}()},
\code{\link{Partition_Pairs}()},
\code{\link{Partition_Type_All}()},
\code{\link{Partition_Type_eL_Location}()}
}
\concept{Partitions}
