% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_All}
\alias{Partition_All}
\title{Generates all disjoint set-partition matrices for N elements}
\usage{
Partition_All(N)
}
\arguments{
\item{N}{The (integer) number of elements to be partitioned}
}
\value{
U the list of partition matrices  of size r.

S_N_r the number of partitions of size r including r=1 blocks,
r=2 blocks etc.
}
\description{
Generates all disjoint set-partition matrices for N elements
}
\examples{
PA<-Partition_All(4)
PA$U
PA$S_N_r
}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 1.4.1 (see Table 1.3, p.23, with partitions for n=4)
}
\seealso{
Other Partitions: 
\code{\link{Partition_2Perm}()},
\code{\link{Partition_Generator}()},
\code{\link{Partition_Indecomposable}()},
\code{\link{Partition_Pairs}()},
\code{\link{Partition_Type_All}()},
\code{\link{Partition_Type_eL_Location}()}
}
\concept{Partitions}
