% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_DiagramsClosedNoLoops}
\alias{Partition_DiagramsClosedNoLoops}
\title{Closed Diagrams without Loops}
\usage{
Partition_DiagramsClosedNoLoops(L)
}
\arguments{
\item{L}{a partition matrix}
}
\value{
The list of partition matrices indecomposable with respect to L
}
\description{
Given a partition L looks for those partitions  \code{K}  of pairs which are 
indecomposable in accordance with L, (L, \code{K}) represent diagrams.
It is used for getting higher order cumulants of Hermite polynomials defined by  L, 
see Terdik (2021) Proposition 4.3, p.194.
}
\references{
Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021. Section 1.4.8.1
}
