% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_Generator}
\alias{Partition_Generator}
\title{Generates all set-partition matrices for N elements}
\usage{
Partition_Generator(N)
}
\arguments{
\item{N}{The (integer) number of elements to be partitioned}
}
\value{
U a matrix containing all the partion matrices

S_N_r the number of partitions of size r including r=1 blocks,
r=2 blocks etc.
}
\description{
Generates all set-partition matrices for N elements
}
\examples{
Partition_Generator(3)
}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 1.4.1
}
\seealso{
Other Partitions: 
\code{\link{Partition_2Perm}()},
\code{\link{Partition_All}()},
\code{\link{Partition_Indecomposable}()},
\code{\link{Partition_Pairs}()},
\code{\link{Partition_Type_All}()},
\code{\link{Partition_Type_eL_Location}()}
}
\concept{Partitions}
