% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_Pairs}
\alias{Partition_Pairs}
\title{Partition into pairs of the set 1:N}
\usage{
Partition_Pairs(N)
}
\arguments{
\item{N}{The (integer) number of elements to be partitioned}
}
\value{
K2 The list of partition matrices, the list
is empty if \code{N} is odd.
}
\description{
Partition into pairs of the set 1:N
}
\references{
Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 1.4.8
}
\seealso{
Other Partitions: 
\code{\link{Partition_2Perm}()},
\code{\link{Partition_All}()},
\code{\link{Partition_Generator}()},
\code{\link{Partition_Indecomposable}()},
\code{\link{Partition_Type_All}()},
\code{\link{Partition_Type_eL_Location}()}
}
\concept{Partitions}
