% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_Type_All}
\alias{Partition_Type_All}
\title{Type and number of partitions}
\usage{
Partition_Type_All(N)
}
\arguments{
\item{N}{The (integer) number of elements to be partitioned}
}
\value{
Part.class A list of all possible partitions

S_N_r A vector with the number of partitions of size r=1, r=2, etc.

eL_r A list with partition types

S_r_j
}
\description{
Generates all partitions of \code{N} numbers and classify by type
}
\examples{
PTA<-Partition_Type_All(5)
# Partitions generated
PTA$Part.class
# Partitions of size 3
PTA$eL_r[[3]]
# Number of partitions with r=1 blocks, r=2 blocks, etc-
PTA$S_N_r
# ????
PTA$S_r_j

}
\references{
Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Definition 1.4
}
\seealso{
Other Partitions: 
\code{\link{Partition_2Perm}()},
\code{\link{Partition_All}()},
\code{\link{Partition_Generator}()},
\code{\link{Partition_Indecomposable}()},
\code{\link{Partition_Pairs}()},
\code{\link{Partition_Type_eL_Location}()}
}
\concept{Partitions}
