% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{Partition_Type_eL_Location}
\alias{Partition_Type_eL_Location}
\title{Find the  parameters  of a given type (eL) of partition}
\usage{
Partition_Type_eL_Location(eL)
}
\arguments{
\item{eL}{A partition of a given type}
}
\value{
r The partition's size

mm The row number in the matrix of types of
that partition
}
\description{
Find the  parameters  of a given type (eL) of partition
}
\examples{
eL<-c(3,1,1,0,0,0,0,0)
Partition_Type_eL_Location(eL)
# eL is in the 5th cell 2nd row

}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 1.4
}
\seealso{
Other Partitions: 
\code{\link{Partition_2Perm}()},
\code{\link{Partition_All}()},
\code{\link{Partition_Generator}()},
\code{\link{Partition_Indecomposable}()},
\code{\link{Partition_Pairs}()},
\code{\link{Partition_Type_All}()}
}
\concept{Partitions}
