% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.r
\name{PermutationInv}
\alias{PermutationInv}
\title{Inverse of a Permutation}
\usage{
PermutationInv(permutation0)
}
\arguments{
\item{permutation0}{A permutation of numbers 1:n}
}
\value{
A vector containing the inverse permutation of  \code{permutation0}
}
\description{
Inverse of a Permutation
}
\references{
Gy. Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021, Remark 1.1, p.2
}
\seealso{
Other Partitions: 
\code{\link{PartitionTypeAll}()},
\code{\link{Partitions}()}
}
\concept{Partitions}
