% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{Rand_CFUSSD_G}
\alias{Rand_CFUSSD_G}
\title{Random multivariate CFUSSD}
\usage{
Rand_CFUSSD_G(n, d, p, a, b, Delta)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}

\item{p}{dimension of the first term}

\item{a}{shape  parameter of the Gamma generator}

\item{b}{scale parameter of the Gamma generator}

\item{Delta}{skewness matrix}
}
\value{
a random matrix \eqn{n x d}
}
\description{
Generate random d-vectors  from the multivariate Canonical Fundamental
Skew-Spherical distribution (CFUSSD)  with Gamma generator
}
\examples{
d <- 2; p <- 3
Lamd <-  matrix(sample(1:50-25, d*p), nrow=d)
ieg<- eigen(diag(p)+t(Lamd)\%*\%Lamd)
V <- ieg$vectors
Delta <-Lamd \%*\% V \%*\% diag(1/sqrt(ieg$values)) \%*\% t(V)
Rand_CFUSSD_G(20,p,d,1,1,Delta)
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021, (5.36) p. 266, (see p.247 for Delta)
}
\seealso{
Other Random generation: 
\code{\link{Rand_CFUSN}()},
\code{\link{Rand_Uni_d}()}
}
\concept{Random generation}
