% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{Rand_SkewNorm}
\alias{Rand_SkewNorm}
\title{Random Multivariate Skew Normal}
\usage{
Rand_SkewNorm(n, Om, alf)
}
\arguments{
\item{n}{sample size}

\item{Om}{correlation matrix   with d dimension}

\item{alf}{shape parameter vector of dimension d}
}
\value{
A random matrix \eqn{n x d}
}
\description{
Generate random d-vectors  from the multivariate Skew Normal distribution
}
\examples{
alpha<-c(10,5,0)
omega<-diag(rep(1,3))
x<-Rand_SkewNorm(20,omega,alpha)

}
\references{
Azzalini, A. with the collaboration of Capitanio, A. (2014).
The Skew-Normal and Related Families. Cambridge University Press,
  IMS Monographs series.

Gy.H.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021, Section 5.1.2
 @family Random generation
}
