% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{Rand_Uni_d}
\alias{Rand_Uni_d}
\title{Random Uniform on the sphere}
\usage{
Rand_Uni_d(n, d)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}
}
\value{
A random matrix \eqn{n x d}
}
\description{
Generate random d-vectors from the Uniform distribution on the sphere
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021

S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644.
}
\seealso{
Other Random generation: 
\code{\link{Rand_CFUSN}()},
\code{\link{Rand_CFUSSD_G}()}
}
\concept{Random generation}
