% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.r
\name{SampleEdg}
\alias{SampleEdg}
\title{Deprecated function}
\usage{
SampleEdg(...)
}
\description{
`SampleEdg()` has been replaced by the more efficient
\code{Edgeworth()} function.
}
\keyword{deprecated}
\keyword{internal}
