% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{SampleEVSK}
\alias{SampleEVSK}
\title{Estimation of multivariate Mean, Variance, T-Skewness and T-Kurtosis vectors}
\usage{
SampleEVSK(X)
}
\arguments{
\item{X}{d-variate data matrix}
}
\value{
The list of the estimated mean, variance, skewness and kurtosis vectors
}
\description{
Provides estimates of mean, variance, skewness and (excess) kurtosis vectors for  d-variate data
}
\examples{
x<- MASS::mvrnorm(100,rep(0,3), 3*diag(rep(1,3)))
EVSK<-SampleEVSK(x)
names(EVSK)
EVSK$estSkew
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021, Sections 6.4.1 and 6.5.1
}
\seealso{
Other Estimation: 
\code{\link{SampleKurt}()},
\code{\link{SampleMomCum}()},
\code{\link{SampleSkew}()},
\code{\link{VarianceKurt}()},
\code{\link{VarianceSkew}()}
}
\concept{Estimation}
