% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.r
\name{SampleHermiteN}
\alias{SampleHermiteN}
\title{Deprecated function}
\usage{
SampleHermiteN(...)
}
\description{
`SampleHermiteN()` has been removed. The \code{HermiteN()} function now
supports both univariate and multivariate Hermite polynomials up to order N.
}
\keyword{deprecated}
\keyword{internal}
