% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{SampleVarianceSkewKurt}
\alias{SampleVarianceSkewKurt}
\title{Estimated Variance of  skewness and kurtosis vectors}
\usage{
SampleVarianceSkewKurt(X)
}
\arguments{
\item{X}{A matrix of d-variate data}
}
\value{
The list of covariance matrices of the skewness and kurtosis vectors
}
\description{
Provides the estimated covariance matrices of the data-estimated skewness and kurtosis
vectors.
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.
}
