% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Stand_Multi}
\alias{Stand_Multi}
\title{Standardize multivariate data}
\usage{
Stand_Multi(x)
}
\arguments{
\item{x}{a multivariate data matrix, sample size is the number of rows}
}
\value{
z multivarate data with null mean vector and
identity sample covariance matrix
}
\description{
For data formed by d-variate vectors x with sample covariance S and sample mean M,
it computes the values
\eqn{z=S^{-1/2}(x-M)}
}
\examples{
x<-rmvnorm(1000,mean=c(0,0,1,3))
z<-Stand_Multi(x)
mu_z<- apply(z,2,mean)
cov_z<- cov(z)

}
\concept{Standard}
