% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.r
\name{Variance_of_Esti_Skew_Th}
\alias{Variance_of_Esti_Skew_Th}
\title{Asymptotic Variance for estimated skewness}
\usage{
Variance_of_Esti_Skew_Th(cum)
}
\arguments{
\item{cum}{The theoretical/estimated cumulants up to order 6 in vector form}
}
\description{
Asymptotic Variance for estimated skewness
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021. skewness (6.13)
}
\seealso{
Other Estimation: 
\code{\link{Esti_MMom_MCum}()},
\code{\link{Esti_MVSK}()},
\code{\link{Esti_Skew_Mardia}()},
\code{\link{Esti_Variance_MVSK}()},
\code{\link{Variance_of_Esti_Kurt_Th}()}
}
\concept{Estimation}
