% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commutators.R
\name{Commutator_Kperm}
\alias{Commutator_Kperm}
\title{Commutator_Kperm}
\usage{
Commutator_Kperm(perm, dims)
}
\arguments{
\item{perm}{vector indicating the permutation of the order #' in the Kronecker product,  use dims <- d if all dimensions #' are equal}

\item{dims}{vector indicating the dimensions of the vectors}
}
\value{
A permutation matrix
}
\description{
Produces any permutation of kronecker products of vectors of any length
Warning: for perm <- c(2,4,6,1,3,8,5,7),
d <- 3, and dims <- c(d,d,d,d,d,d,d,d)
 time: 1326.47 sec elapsed if  dims <- 3, then 2.11 sec  #'elapsed, faster!!
}
\examples{
dims=c(2,3,2)
perm = c(1,3,2)
K.matr(perm,dims)
perm =c(3,1,4,2) dims=4 # All vectors with dimension 4
}
\references{
Holmquist B (1996) The d-variate vector Hermite polynomial of order. Linear Algebra
Appl 237/238:155–190

Gy., Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021, 1.2.4 Commuting T-Products of Vectors.
}
