% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commutators.R
\name{Commutator_Moment_L}
\alias{Commutator_Moment_L}
\title{Commutator_Moment_L}
\usage{
Commutator_Moment_L(el_rm, r, m, d)
}
\description{
Commutator_Moment_L
}
\examples{
n=4
r=2
m=1
d=2
PTA<-Partition_Type_All(n)
el_rm<-PTA$eL_r[[2]][1,]
el_r_m is always a vector
}
\references{
Gy., Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021, 2.3, Corollary 2.6.
}
