% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomCums.r
\name{Cum2Mom}
\alias{Cum2Mom}
\title{Get moments from  cumulants (univariate)}
\usage{
Cum2Mom(cum_x)
}
\arguments{
\item{cum_x}{the n-vector of cumulants starting from the fist - the mean
and arriving to the n-th order cumulant}
}
\value{
mu_x the vector of univariate moments
}
\description{
Obtains a vector of univariate moments from a vector of
univariate cumulants
}
\examples{
cum_x<- c(1,2,3,4,5,6,7,8)
Cum2Mom(cum_x)

}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 3.4 formula 3.23
}
\seealso{
Other Moments and cumulants: 
\code{\link{Cum2MomMulti}()},
\code{\link{Mom2CumMulti}()},
\code{\link{Mom2Cum}()}
}
\concept{Moments and cumulants}
