% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_Coeff}
\alias{Hermite_Coeff}
\title{Hermite_Coeff
n the Hermite polynomial Hen(x) of variance 1, 
the absolute value of the #'coefficient of \eqn{x^k} 
is the number of (unordered) partitions of an n-element
set into \eqn{k} singletons and \eqn{(n ??? k)/2}   (unordered) pairs.
coefficient of \eqn{x^k} in  \eqn{H_N(x)}}
\usage{
Hermite_Coeff(N)
}
\description{
Hermite_Coeff
n the Hermite polynomial Hen(x) of variance 1, 
the absolute value of the #'coefficient of \eqn{x^k} 
is the number of (unordered) partitions of an n-element
set into \eqn{k} singletons and \eqn{(n ??? k)/2}   (unordered) pairs.
coefficient of \eqn{x^k} in  \eqn{H_N(x)}
}
\seealso{
Other Hermite: 
\code{\link{Hermite_CoeffMulti}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Inv}()},
\code{\link{Hermite_Poly_NH_Multi_Inv}()}
}
\concept{Hermite}
