% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_CoeffMulti}
\alias{Hermite_CoeffMulti}
\title{Coefficients of multivariate T-Hermite polynomials for standardized variate}
\usage{
Hermite_CoeffMulti(N, d)
}
\arguments{
\item{N}{the maximum order of polynomials}

\item{d}{the dimension of  d-variate X}
}
\value{
The list of matrices of coefficients for the d-variate polynomials from 1 to N
}
\description{
Provides the matrix of coefficients of
\eqn{x^{\otimes N}}, \eqn{x^{\otimes (N-2)}}...
for the d-variate T-Hermite polynomials  up to order N.
}
\references{
Gy.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 4.6.2, p. 223, use Property 4.10 of multilinearity for general case
}
\seealso{
Other Hermite: 
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_N_Cov_X1_X2}()},
\code{\link{Hermite_Nth}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Inv}()},
\code{\link{Hermite_Poly_NH_Multi_Inv}()}
}
\concept{Hermite}
