% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hermite.r
\name{Hermite_Poly_NH_Multi_Inv}
\alias{Hermite_Poly_NH_Multi_Inv}
\title{Inverse d-variate Hermite Polynomial}
\usage{
Hermite_Poly_NH_Multi_Inv(H_N_X, Sig2, N)
}
\description{
Inverse d-variate Hermite Polynomial
}
\examples{
Sig2=matrix(c(1,0,0,1),2,2,byrow=T)
x<-c(1,3)
N<-4
H_N_X<-Hermite_Poly_HN_Multi(x,sigma2,N)
Hermite_Poly_NH_Multi_Inv(H_N_X,Sig2,N)
}
\seealso{
Other Hermite: 
\code{\link{Hermite_CoeffMulti}()},
\code{\link{Hermite_Coeff}()},
\code{\link{Hermite_Poly_HN_Multi}()},
\code{\link{Hermite_Poly_HN}()},
\code{\link{Hermite_Poly_NH_Inv}()}
}
\concept{Hermite}
