% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomCums.r
\name{Mom2Cum}
\alias{Mom2Cum}
\title{Get cumulants from moments (univariate)}
\usage{
Mom2Cum(mu_x)
}
\arguments{
\item{mu_x}{the n-vector of moments starting from the fist moment - the mean
and arriving to the n-th order moment}
}
\value{
cum_x the vector of univariate cumulants
}
\description{
Obtains a vector of univariate cumulants from a vector of
univariate moments
}
\examples{
mu_x<- c(1,2,3,4,5,6,7,8)
Mom2Cum(mu_x)

}
\references{
G.H.Terdik, Multivariate statistical methods - going beyond the linear,
Springer 2021.  Section 3.4 formula 3.20
}
\seealso{
Other Moments and cumulants: 
\code{\link{Cum2MomMulti}()},
\code{\link{Cum2Mom}()},
\code{\link{Mom2CumMulti}()}
}
\concept{Moments and cumulants}
