% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{rUniS}
\alias{rUniS}
\title{Random Uniform on the sphere}
\usage{
rUniS(n, d)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}
}
\value{
A random matrix \eqn{n \times d}
}
\description{
Generate random d-vectors from the Uniform distribution on the sphere
}
\references{
Gy.Terdik, Multivariate statistical methods - Going beyond the linear,
Springer 2021

S. R. Jammalamadaka, E. Taufer, Gy. Terdik. On multivariate
skewness and kurtosis. Sankhya A, 83(2), 607-644.
}
\seealso{
Other Random generation: 
\code{\link{rCFUSN}()},
\code{\link{rCFUSSD}()},
\code{\link{rSkewNorm}()}
}
\concept{Random generation}
